+++ src/types/poll_option.rs
@@ -6,5 +6,5 @@ pub struct PollOption {
     /// Option text, 1-100 characters
     pub text: String,
     /// Number of users that voted for this option
-    pub voter_count: u32,
+    pub voter_count: i64,
 }
